fix: convert `coolify_server` to new model
